pyproject.toml
[tool.poetry]
name = "translationservice"
version = "0.1.0"
description = ""
authors = ["Jonathen Sundy <jonathensundy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.14"
pandas = "^2.2.2"
datasets = "^2.19.2"
azure-ai-translation-text = "^1.0.0"
tqdm = "^4.66.4"
load-dotenv = "^0.1.0"
mock = "^5.1.0"
huggingface-hub = "^0.23.3"
transformers = "^4.41.2"
sentencepiece = "^0.2.0"
sacremoses = "^0.1.1"
torch = "^2.3.1"
boto3 = "^1.34.124"
langchain = "^0.2.5"
openai = "^1.34.0"
ollama = "^0.2.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
